Move FeatureListNamespaceModel to feature_list_namespace.py
